requirements.txt
cocos2d
numpy
pillow
pygame
pyglet
pyyaml
